rebase

+++ test/evm/hardhat/fixtures/SpokePool.Fixture.ts
@@ -446,7 +446,7 @@ export async function getUpdatedV3DepositSignature(
   const typedData = {
     types: {
       UpdateDepositDetails: [
-        { name: "depositId", type: "uint256" },
+        { name: "depositId", type: "uint32" },
         { name: "originChainId", type: "uint256" },
         { name: "updatedOutputAmount", type: "uint256" },
         { name: "updatedRecipient", type: "address" },
